Refactor router module
