+++ server/core/__init__.py
@@ -1,6 +1,6 @@
 from .dependency_injector import DependencyInjector
 from .protocol import Protocol
-from .routing import RouteError, Router
+from .router import RouteError, Router
 from .service import Service, create_services
 
 __all__ = (
