--- server/core/routing/__init__.py
@@ -1,6 +1,0 @@
-from .router import RouteError, Router, SearchTree
-
-__all__ = (
-    "RouteError"
-    "Router",
-)
