+++ server/core/typedefs.py
@@ -1,0 +1,6 @@
+from typing import Any, Callable, Dict
+
+# Type aliases
+Handler = Callable[..., Any]
+Message = Dict[Any, Any]
+
